renamed to AuditLogger
